bitbake: server/process: Avoid hanging if a parser process is terminated
If a parser process is terminated while holding a write lock, then it
will lead to a deadlock (see
https://docs.python.org/3/library/multiprocessing.html#multiprocessing.Process.terminate).

With SIGTERM, we don't want to terminate holding the lock. We also don't
want a SIGINT to cause a partial write to the event stream.

I tried using signal masks to avoid this but it doesn't work, see
https://bugs.python.org/issue47139

Instead, add a signal handler and catch the calls around the critical section
which is heavier but safe and seems to avoid all the deadlocks I was able to
trigger.

Some ideas from Peter Kjellerstedt <peter.kjellerstedt@axis.com>

(Bitbake rev: ce0dc518bbb0d36b4be19dc8292818168c9cc181)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>